Fixed some try catch blocks, improved the logic of file not found on s3

+++ backend/src/server.ts
@@ -9,6 +9,7 @@ import {
 import { addFile, deleteFile, getAllFiles, getFile } from "./utils/sqlite";
 import {
   deleteFromS3,
+  doesObjectExist,
   getSignedDownloadURL,
   isDeleteSuccess,
   isUploadSuccess,
@@ -30,7 +31,7 @@ apiRouter.post(
   "/upload",
 
   singleFileHandler, // handle the file upload using multer
-  (req, res, next) => {
+  (_req, _res, next) => {
     // basically checking if the file is uploaded
     /*
     console.log(`req body is `, req.body);
@@ -47,41 +48,51 @@ apiRouter.post(
     next();
   },
   async (req, res) => {
-    // upload the file to s3 and save the file to the sqlite db
-    if (!req.file) {
-      res.status(400).json({ error: "No file uploaded" });
-      return;
-    }
-    // temproray commented
-    const fileUploadToS3 = await uploadToS3(req.file?.path, req.file?.filename);
-    if (!isUploadSuccess(fileUploadToS3)) {
-      res.status(400).json({ error: "File upload failed to s3" });
-      return;
-    }
+    try {
+      // upload the file to s3 and save the file to the sqlite db
+      if (!req.file) {
+        res.status(400).json({ error: "No file uploaded" });
+        return;
+      }
+      // temproray commented
+      const fileUploadToS3 = await uploadToS3(
+        req.file?.path,
+        req.file?.filename
+      );
+      if (!isUploadSuccess(fileUploadToS3)) {
+        res.status(400).json({ error: "File upload failed to s3" });
+        return;
+      }
 
-    // lets save to the sqlite db
-    const fileID = await addFile(req.file.originalname, req.file.filename);
-    if (!fileID) {
-      res.status(400).json({ error: "File upload failed to sqlite" });
-      return;
+      // lets save to the sqlite db
+      const fileID = await addFile(req.file.originalname, req.file.filename);
+      if (!fileID) {
+        res.status(400).json({ error: "File upload failed to sqlite" });
+        return;
+      }
+
+      // Lets delete the local file in the server
+      // console.log(`\n \n we need to delete this file `, req.file);
+      const isFileDeleted = clearTempFiles(req.file.path);
+      console.log("local file copy deleted successfully", isFileDeleted);
+
+      // if (!isFileDeleted) {
+      //   res.status(400).json({ error: "File upload failed to sqlite" });
+      //   return;
+      // }
+      res.status(200).json({
+        downloadUrl: `${
+          process.env.VITE_DOMAIN_NAME || "http//localhost:4001"
+        }/d/${fileID}`,
+        message: "File uploaded successfully",
+      });
+      // FUTURE WORK ⚒️
+      // Now we send the url through which the user can share the file.
+      // On the spot we will have some api to turn this url into some QR code
+    } catch (err: any) {
+      console.error("Some weird error ", err);
+      res.status(500).json({ error: "Internal server error" });
     }
-
-    // Lets delete the local file in the server
-    // console.log(`\n \n we need to delete this file `, req.file);
-    const isFileDeleted = clearTempFiles(req.file.path);
-    console.log("local file copy deleted successfully", isFileDeleted);
-
-    // if (!isFileDeleted) {
-    //   res.status(400).json({ error: "File upload failed to sqlite" });
-    //   return;
-    // }
-    res.status(200).json({
-      downloadUrl: `${process.env.VITE_DOMAIN_NAME||"http//localhost:4001"}/d/${fileID}`,
-      message: "File uploaded successfully",
-    });
-    // FUTURE WORK ⚒️
-    // Now we send the url through which the user can share the file.
-    // On the spot we will have some api to turn this url into some QR code
   }
 );
 
@@ -97,27 +108,43 @@ apiRouter.get("/download/:location", (req, res) => {
 });
 
 // To see all the files that are uploaded
-apiRouter.get("/dall", async (req, res) => {
-  const files = await getAllFiles();
-  res.status(200).json({ files });
+apiRouter.get("/dall", async (_req, res) => {
+  try {
+    const files = await getAllFiles();
+    res.status(200).json({ files });
+  } catch (error: any) {
+    console.log(error);
+    res.status(500).json({ error: "Internal server error" });
+  }
 });
 
 apiRouter.get("/del/:id", async (req, res) => {
   const id = req.params.id;
   console.log("The delete id is ", id);
-  const s3FilePath = (await getFile(id)).s3path;
-  const deleteResponse = await deleteFromS3(s3FilePath);
-  if (deleteResponse===false)
-    res.status(400).json({ error: "File deletion failed, not found" });
-  else if (isDeleteSuccess(deleteResponse)) {
-    const isFileDeleted = await deleteFile(id);
-    if (isFileDeleted === null) {
-      console.log('The delete was successful in file db');
-      res.status(200).json({ message: "File deleted successfully" });
-    } else res.status(400).json({ error: "File deletion failed" });
-  } else {
-    res.status(400).json({ error: "File deletion failed" });
-  }
+  getFile(id)
+    .then(async (existingFile) => {
+      if (!existingFile) {
+        res.status(404).json({ error: "File not found" });
+        return;
+      }
+      const s3FilePath = existingFile.s3path;
+      const deleteResponse = await deleteFromS3(s3FilePath);
+      if (deleteResponse === false)
+        res.status(400).json({ error: "File deletion failed, not found" });
+      else if (isDeleteSuccess(deleteResponse)) {
+        const isFileDeleted = await deleteFile(id);
+        if (isFileDeleted === null) {
+          console.log("The delete was successful in file db");
+          res.status(200).json({ message: "File deleted successfully" });
+        } else res.status(400).json({ error: "File deletion failed" });
+      } else {
+        res.status(400).json({ error: "File deletion failed" });
+      }
+    })
+    .catch((_err) => {
+      res.status(400).json({ error: "File not found" });
+      return;
+    });
 });
 
 // The actual implementation to send the filename and the url
@@ -126,18 +153,36 @@ apiRouter.get("/d/:id", async (req, res) => {
   // make a request to 4001/api/v0/download/path and get the file name and the url
 
   const id = req.params.id;
-  const file = await getFile(id);
-  console.log("file info in the /d is ", file);
+  getFile(id)
+    .then(async (file) => {
+      console.log("file info in the /d is ", file);
 
-  const downloadUrl = await getSignedDownloadURL(file?.s3path, file.file_name);
-  if (!downloadUrl) {
-    res.status(400).json({ error: "File download failed" });
-    return;
-  }
-  res.status(200).json({
-    filename: file?.file_name,
-    downloadUrl: downloadUrl,
-  });
+      // ⚠️ check before geting the download url, if file exists on the s3
+      const isFileExists = await doesObjectExist(file?.s3path);
+      if (!isFileExists) {
+        // ⚠️ remove this entry from the db too
+        await deleteFile(file.id);
+        res.status(400).json({ error: "File not found on the s3 bucket" });
+        return;
+      }
+      const downloadUrl = await getSignedDownloadURL(
+        file?.s3path,
+        file.file_name
+      );
+      if (!downloadUrl) {
+        res.status(400).json({ error: "File download failed" });
+        return;
+      }
+      res.status(200).json({
+        filename: file?.file_name,
+        downloadUrl: downloadUrl,
+      });
+    })
+    .catch((err: Error) => {
+      console.error("error in the /d/:id route", err);
+      res.status(400).json({ error: "File not found" });
+      return;
+    });
 });
 
 // the base api route
